Retrieve and Enhance Building Data with RegBl data and CO2 Emissions Calculations

This function enriches a given building's data record by fetching additional details from the RegBl API, converting the data to the SIA format used by the co2calculatorPACTA2022 package, and computing the building's CO2 emissions. If the building has been demolished, as indicated by a non-NA GABBJ value, the function halts and reports the demolition. Otherwise, it proceeds to convert the format of the building data and calculate emissions using various building attributes.

Usage

fill_building_data(building, sqlite_conn = NULL)

Arguments

building

A list or data frame containing at least minimal identifying information for a building, which is then used to fetch and augment data from the RegBl database.

data_source

A character string specifying the source of the RegBl data, either 'sqlite' or 'web'.

Value

An enhanced version of the input building data, including additional details and computed CO2 emissions.

Examples

building_data <- list(EGID = "123456")
enriched_building_data <- fill_building_data(building_data)
#> Error in fill_building_data(building_data): could not find function "fill_building_data"