Fill missing data in the buildings dataframe with RegBl data and computed CO2 emissions

This function processes a dataframe of buildings, applying the fill_building_data function to each row/building in parallel. It ensures that the input dataframe contains all necessary columns, filling any missing ones with NA values, then computes detailed building information and CO2 emissions for each building. The function leverages parallel computing capabilities to enhance performance and efficiency.

Usage

fill_buildings_df(buildings_df, regbl_db_path = NULL)

Arguments

buildings_df

A dataframe where each row represents a building and contains minimal data needed for enhancement, such as an identifier used to fetch additional data.

Value

A new dataframe with the same structure as buildings_df but augmented with detailed building data and CO2 emissions computations for each building. If any data cannot be retrieved or calculated, the corresponding building will have its error_comments field updated with the error message.

Details

The function first standardizes the input dataframe to ensure it includes all necessary columns, leveraging the standardize_buildings_df function. Then, it uses the future.apply::future_lapply function to apply fill_building_data to each building in the dataframe in parallel, improving processing time on multicore systems. Each building's data is individually retrieved and processed, with results aggregated into a single dataframe returned at the end.

The parallel execution model can be adjusted based on system resources by changing the number of workers in the future::plan call.

Examples

buildings_df <- data.frame(EGID = c("123456", "789012"))
enhanced_buildings_df <- fill_buildings_df(buildings_df)
#> Error in globalCallingHandlers(condition = global_progression_handler): should not be called with handlers on the stack